+++ src/ConfigCatClient/Cache/CacheParameters.cs
@@ -2,9 +2,7 @@ namespace ConfigCat.Client.Cache;
 
 internal class CacheParameters
 {
-#pragma warning disable CS0618 // Type or member is obsolete
-    public IConfigCache ConfigCache { get; set; } // Backward compatibility, it'll be changed to IConfigCatCache later.
-#pragma warning restore CS0618 // Type or member is obsolete
+    public IConfigCatCache ConfigCache { get; set; }
 
     public string CacheKey { get; set; }
 }
