+++ src/ConfigCatClient/Cache/IConfigCatCache.cs
@@ -1,16 +1,12 @@
+using System.Threading.Tasks;
+using System.Threading;
+
 namespace ConfigCat.Client;
 
 /// <summary>
 /// Defines a cache used by the <see cref="ConfigCatClient"/>.
 /// </summary>
-/// <remarks>
-/// Note for implementers. Until the deprecated <see cref="IConfigCache"/> interface is removed, this interface needs to extend it for backward compatibility.
-/// Later, all of its members will be moved into this interface.
-/// </remarks>
-public interface IConfigCatCache :
-#pragma warning disable CS0618 // Type or member is obsolete
-    IConfigCache
-#pragma warning restore CS0618 // Type or member is obsolete
+public interface IConfigCatCache
 {
     /// <summary>
     /// Sets a <see cref="ProjectConfig"/> into cache.
@@ -20,8 +16,24 @@ public interface IConfigCatCache :
     void Set(string key, ProjectConfig config);
 
     /// <summary>
+    /// Sets a <see cref="ProjectConfig"/> into cache.
+    /// </summary>
+    /// <param name="key">A string identifying the <see cref="ProjectConfig"/> value.</param>
+    /// <param name="config">The config to cache.</param>
+    Task SetAsync(string key, ProjectConfig config);
+
+    /// <summary>
     /// Gets a <see cref="ProjectConfig"/> from cache.
     /// </summary>
-    /// <returns>The cached config.</returns>
+    /// <param name="key">A string identifying the <see cref="ProjectConfig"/> value.</param>
+    /// <returns>The cached config or <see cref="ProjectConfig.Empty"/> if there is none.</returns>
     ProjectConfig Get(string key);
+
+    /// <summary>
+    /// Gets a <see cref="ProjectConfig"/> from cache.
+    /// </summary>
+    /// <param name="key">A string identifying the <see cref="ProjectConfig"/> value.</param>
+    /// <param name="cancellationToken">An optional <see cref="CancellationToken"/>.</param>
+    /// <returns>The cached config or <see cref="ProjectConfig.Empty"/> if there is none.</returns>
+    Task<ProjectConfig> GetAsync(string key, CancellationToken cancellationToken = default);
 }
