+++ src/ConfigCatClient/ConfigService/LazyLoadConfigService.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Threading.Tasks;
 using ConfigCat.Client.Cache;
-using ConfigCat.Client.Utils;
 
 namespace ConfigCat.Client.ConfigService;
 
@@ -19,30 +18,23 @@ internal sealed class LazyLoadConfigService : ConfigServiceBase, IConfigService
 
     public ProjectConfig GetConfig()
     {
-        // check for the new cache interface until we remove the old IConfigCache.
-        if (this.ConfigCache is IConfigCatCache cache)
+        var config = this.ConfigCache.Get(base.CacheKey);
+
+        if (config.IsExpired(expiration: this.cacheTimeToLive, out var cachedConfigIsEmpty))
         {
-            var config = cache.Get(base.CacheKey);
-
-            if (config.IsExpired(expiration: this.cacheTimeToLive, out var cachedConfigIsEmpty))
+            if (!cachedConfigIsEmpty)
             {
-                if (!cachedConfigIsEmpty)
-                {
-                    OnConfigExpired();
-                }
-
-                if (!IsOffline)
-                {
-                    var configWithFetchResult = RefreshConfigCore(config);
-                    return configWithFetchResult.Item1;
-                }
+                OnConfigExpired();
             }
 
-            return config;
+            if (!IsOffline)
+            {
+                var configWithFetchResult = RefreshConfigCore(config);
+                return configWithFetchResult.Item1;
+            }
         }
 
-        // worst scenario, fallback to sync over async, delete when we enforce IConfigCatCache.
-        return Syncer.Sync(GetConfigAsync);
+        return config;
     }
 
     public async Task<ProjectConfig> GetConfigAsync()
