+++ src/ConfigCatClient/Evaluation/IRolloutEvaluator.cs
@@ -4,7 +4,4 @@ internal interface IRolloutEvaluator
 {
     EvaluationDetails Evaluate(Setting setting, string key, string logDefaultValue, User user,
         ProjectConfig remoteConfig, EvaluationDetailsFactory detailsFactory);
-
-    EvaluationDetails EvaluateVariationId(Setting setting, string key, string logDefaultVariationId, User user,
-        ProjectConfig remoteConfig, EvaluationDetailsFactory detailsFactory);
 }
