+++ src/ConfigCatClient/Logging/IConfigCatLogger.cs
@@ -5,17 +5,13 @@ namespace ConfigCat.Client;
 /// <summary>
 /// Defines the interface for the ConfigCat SDK to perform logging.
 /// </summary>
-/// <remarks>
-/// Note for implementers. Until the deprecated <see cref="ILogger"/> interface is removed, this interface needs to extend it for backward compatibility.
-/// This means that when implementing this interface, you also need to implement the full <see cref="ILogger"/> interface temporarily.
-/// Later, the <see cref="ILogger.LogLevel"/> property will be moved into this interface but the other methods like <see cref="ILogger.Error(string)"/> will be removed.
-/// (However, the ConfigCat SDK does not use the old methods internally any more, so you do not need to provide an actual implementation for them.)
-/// </remarks>
-public interface IConfigCatLogger :
-#pragma warning disable CS0618 // Type or member is obsolete
-    ILogger
-#pragma warning restore CS0618 // Type or member is obsolete
+public interface IConfigCatLogger
 {
+    /// <summary>
+    /// Specifies message filtering.
+    /// </summary>
+    LogLevel LogLevel { get; set; }
+
     /// <summary>
     /// Writes a message into the log.
     /// </summary>
